Create a metadata store interface
